Timezones internals support -- understand manually expanded "M" format.
Currently PST8PDT,M3.2.0/2:00:00,M11.1.0/2:00:00 is hardcoded as tz2,
however internals should work for all timezones that have Posix "M" format.
This implementation is intended for a watch face, so it does not need
any historical data. It should be able to show time that is right now,
with a given timezone format. "J" and "j" formats are not implemented yet.